moving UI components with self-contained state to their own dir for clarification
