+++ app/views/components/disk-allocator/index.js
