+++ app/views/components/external-anchor/index.js
