+++ app/views/components/modal/index.js
