+++ app/views/components/notification/index.js
