+++ app/views/share-wizard/wizard0.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   components: {
-    'ext-a' : require('../external-anchor')
+    'ext-a' : require('../components/external-anchor')
   },
   template: `
 <section>
