+++ app/views/terms/content.js
@@ -6,7 +6,7 @@
 
 module.exports = {
   components: {
-    'ext-a': require('../external-anchor')
+    'ext-a': require('../components/external-anchor')
   },
   template: `
 <section>
