+++ app/views/updater/index.js
@@ -9,7 +9,7 @@ const updater = require('../../lib/updater');
 
 module.exports = {
   components: {
-    'modal': require('../modal')
+    'modal': require('../components/modal')
   },
   data: function() {
     return {
